requirements.txt
# ============================================================
# RSS FEED PROCESSING - Core Dependencies
# ============================================================
feedparser>=6.0.10          # Parsing RSS/Atom feeds
requests>=2.31.0            # HTTP requests
aiohttp>=3.9.0              # Async HTTP requests
beautifulsoup4>=4.12.0      # HTML parsing
lxml>=4.9.0                 # XML/HTML parser (faster than html.parser)
python-dateutil>=2.8.2      # Date parsing utilities

# ============================================================
# ARTICLE EXTRACTION & PROCESSING
# ============================================================
trafilatura>=1.6.0          # Main text extraction from articles
newspaper3k>=0.2.8          # Alternative article extraction
readability-lxml>=0.8.1     # Article content extraction
langdetect>=1.0.9           # Language detection

# ============================================================
# DATA VALIDATION & MODELS
# ============================================================
pydantic>=2.5.0             # Data validation and settings management

# ============================================================
# CONFIGURATION & UTILITIES
# ============================================================
PyYAML>=6.0.1               # YAML configuration files
tenacity>=8.2.0             # Retry logic with exponential backoff
tqdm>=4.66.0                # Progress bars

# ============================================================
# PYTHON COMPATIBILITY & SYSTEM DEPENDENCIES
# ============================================================
setuptools>=65.0.0          # Required for pkg_resources (fixes ModuleNotFoundError)
importlib-metadata>=4.0.0   # Metadata for installed packages (Python < 3.10 compatibility)

# ============================================================
# GUI (Tkinter viene con Python, pero necesitamos PIL)
# ============================================================
Pillow>=10.0.0              # Image processing for GUI

# ============================================================
# BROWSER AUTOMATION (Optional - for blocked sites)
# ============================================================
# Uncomment if needed for sites that block scrapers
# playwright>=1.40.0
# selenium>=4.15.0

# ============================================================
# DEVELOPMENT & TESTING (Optional)
# ============================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0

# ============================================================
# LANGCHAIN & LLM INTEGRATION
# ============================================================
langchain>=0.1.0              # Core LangChain framework
langchain-groq>=0.0.1         # Groq integration for LangChain
langchain-core>=0.1.0         # LangChain core components
python-dotenv>=1.0.0          # Environment variable management
openpyxl>=3.1.0               # Excel export support